requirements.txt
# Existing dependencies
psutil>=5.9.0
anthropic>=0.45.0
pytest>=8.0.0
requests>=2.32.3
python-dotenv>=1.0.0
beautifulsoup4>=4.12.3
pyautogui>=0.9.54

# Additional core dependencies
openai>=1.0.0  # For Deepseek API compatibility

# Additional testing
pytest-asyncio>=0.23.0  # For async test support
responses>=0.24.0  # For HTTP mocking

# Additional computer tool dependencies
pywin32>=306; platform_system=="Windows"  # For Windows-specific functionality

# Additional utilities
typing-extensions>=4.5.0  # For type hints

# Development tools
black>=23.0.0  # Code formatting
isort>=5.12.0  # Import sorting
mypy>=1.5.0  # Type checking

# New dependencies
responses>=0.24.0  # For responses package 
transformers>=4.37.2  # For transformers package
sentencepiece>=0.1.99  # Add this for the tokenizer
torch>=2.0.0  # Add this for the tokenizer
watchdog==3.0.0  # For file watching
protobuf>=4.25.1  # Required for tokenizer functionality
ollama